(BSR)[API] fix: remove repository/providable_queries.py
